+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2020-02-20
+### Added
+- document_oauth2_authentication function to document OAuth2 security definition.
+- document_endpoint_oauth2_authentication function to document OAuth2 security on an endpoint.
+
 ## [0.0.2] - 2020-02-19
 ### Added
 - Set basePath to X-Forwarded-Prefix header if provided.
@@ -14,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/Colin-b/apispec_starlette/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/Colin-b/apispec_starlette/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/Colin-b/apispec_starlette/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/Colin-b/apispec_starlette/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Colin-b/apispec_starlette/releases/tag/v0.0.1